converted the permissions module to use the action controllers
