+++ ACP3/Modules/ACP3/Permissions/Controller/Admin/Index/Create.php
@@ -1,0 +1,144 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Permissions\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Permissions;
+
+/**
+ * Class Create
+ * @package ACP3\Modules\ACP3\Permissions\Controller\Admin\Index
+ */
+class Create extends AbstractFormAction
+{
+    /**
+     * @var \ACP3\Core\NestedSet
+     */
+    protected $nestedSet;
+    /**
+     * @var \ACP3\Core\Helpers\FormToken
+     */
+    protected $formTokenHelper;
+    /**
+     * @var \ACP3\Modules\ACP3\Permissions\Cache
+     */
+    protected $permissionsCache;
+    /**
+     * @var \ACP3\Modules\ACP3\Permissions\Validation\RoleFormValidation
+     */
+    protected $roleFormValidation;
+
+    /**
+     * Create constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext                   $context
+     * @param \ACP3\Core\NestedSet                                         $nestedSet
+     * @param \ACP3\Core\Helpers\FormToken                                 $formTokenHelper
+     * @param \ACP3\Modules\ACP3\Permissions\Model\RuleRepository          $ruleRepository
+     * @param \ACP3\Modules\ACP3\Permissions\Cache                         $permissionsCache
+     * @param \ACP3\Modules\ACP3\Permissions\Validation\RoleFormValidation $roleFormValidation
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Permissions\Model\RuleRepository $ruleRepository,
+        Core\NestedSet $nestedSet,
+        Core\Helpers\FormToken $formTokenHelper,
+        Permissions\Cache $permissionsCache,
+        Permissions\Validation\RoleFormValidation $roleFormValidation
+    )
+    {
+        parent::__construct($context, $ruleRepository);
+
+        $this->nestedSet = $nestedSet;
+        $this->formTokenHelper = $formTokenHelper;
+        $this->permissionsCache = $permissionsCache;
+        $this->roleFormValidation = $roleFormValidation;
+    }
+
+    /**
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    public function execute()
+    {
+        if ($this->request->getPost()->isEmpty() === false) {
+            return $this->executePost($this->request->getPost()->all());
+        }
+
+        $this->formTokenHelper->generateFormToken();
+
+        return [
+            'modules' => $this->fetchModulePermissions(0, 2),
+            'parent' => $this->fetchRoles(),
+            'form' => array_merge(['name' => ''], $this->request->getPost()->all())
+        ];
+    }
+
+    /**
+     * @param array $formData
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     * @throws \Doctrine\DBAL\ConnectionException
+     */
+    protected function executePost(array $formData)
+    {
+        return $this->actionHelper->handleCreatePostAction(function () use ($formData) {
+            $this->roleFormValidation->validate($formData);
+
+            $insertValues = [
+                'id' => '',
+                'name' => Core\Functions::strEncode($formData['name']),
+                'parent_id' => $formData['parent_id'],
+            ];
+
+            $roleId = $this->nestedSet->insertNode(
+                (int)$formData['parent_id'],
+                $insertValues,
+                Permissions\Model\RoleRepository::TABLE_NAME,
+                true
+            );
+
+            $this->saveRules($formData['privileges'], $roleId);
+
+            $this->permissionsCache->saveRolesCache();
+
+            $this->formTokenHelper->unsetFormToken();
+
+            return $roleId;
+        });
+    }
+
+    /**
+     * @param int $roleId
+     * @param int $moduleId
+     * @param int $privilegeId
+     * @param int $defaultValue
+     *
+     * @return array
+     */
+    protected function generatePrivilegeCheckboxes($roleId, $moduleId, $privilegeId, $defaultValue)
+    {
+        $permissions = [
+            0 => 'deny_access',
+            1 => 'allow_access',
+            2 => 'inherit_access'
+        ];
+
+        $select = [];
+        foreach ($permissions as $value => $phrase) {
+            if ($roleId === 1 && $value === 2) {
+                continue;
+            }
+
+            $select[$value] = [
+                'value' => $value,
+                'selected' => $this->privilegeIsChecked($moduleId, $privilegeId, $value, $defaultValue),
+                'lang' => $this->translator->t('permissions', $phrase)
+            ];
+        }
+
+        return $select;
+    }
+}
